+++ cmd/dossierx/surface_test.go
@@ -1186,7 +1186,6 @@ func surfacePayloadTypes() map[string]any {
 		"lockData":              lockData{},
 		"policyLockPreviewData": policyLockPreviewData{},
 		"policyLockData":        policyLockData{},
-		"policyMigrationData":   policyMigrationData{},
 		"recoverData":           recoverData{},
 		"lockRefusedData":       lockRefusedData{},
 		"batchLockData":         batchLockData{},